# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Accounting & Business Management API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /sale-invoices:
    get:
      operationId: listSaleInvoices
      tags: [saleInvoices]
      summary: List sale invoices
      security:
        - bearerAuth: []
      parameters:
        - name: from
          in: query
          schema:
            type: string
        - name: to
          in: query
          schema:
            type: string
        - name: partyId
          in: query
          schema:
            type: integer
        - name: search
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Sale invoices
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SaleInvoice"
    post:
      operationId: createSaleInvoice
      tags: [saleInvoices]
      summary: Create sale invoice
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSaleInvoiceBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SaleInvoice"
  /sale-invoices/{id}:
    get:
      operationId: getSaleInvoice
      tags: [saleInvoices]
      summary: Get sale invoice
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Invoice
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SaleInvoice"
        "404":
          description: Not found
    patch:
      operationId: updateSaleInvoice
      tags: [saleInvoices]
      summary: Update sale invoice
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateSaleInvoiceBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SaleInvoice"
    delete:
      operationId: deleteSaleInvoice
      tags: [saleInvoices]
      summary: Delete sale invoice
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ORDERS
...
  /reports/day-book:
    get:
      operationId: getDayBook
      tags: [reports]
      summary: Day book report
      security:
        - bearerAuth: []
      parameters:
        - name: date
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Day book entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DayBookEntry"
  /reports/trial-balance:
    get:
      operationId: getTrialBalance
      tags: [reports]
      summary: Trial balance
      security:
        - bearerAuth: []
      parameters:
        - name: asOf
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Trial balance
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TrialBalanceRow"
  /reports/profit-loss:
    get:
      operationId: getProfitLoss
      tags: [reports]
      summary: Profit and loss statement
      security:
        - bearerAuth: []
      parameters:
        - name: from
          in: query
          required: true
          schema:
            type: string
        - name: to
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: P&L report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProfitLoss"
  /reports/balance-sheet:
    get:
      operationId: getBalanceSheet
      tags: [reports]
      summary: Balance sheet
      security:
        - bearerAuth: []
      parameters:
        - name: asOf
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Balance sheet
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BalanceSheet"
  /reports/sale-register:
    get:
      operationId: getSaleRegister
      tags: [reports]
      summary: Sale register
      security:
        - bearerAuth: []
      parameters:
        - name: from
          in: query
          required: true
          schema:
            type: string
        - name: to
          in: query
          required: true
          schema:
            type: string
        - name: partyId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Sale register
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SaleRegisterRow"
  /reports/purchase-register:
    get:
      operationId: getPurchaseRegister
      tags: [reports]
      summary: Purchase register
      security:
        - bearerAuth: []
      parameters:
        - name: from
          in: query
          required: true
          schema:
            type: string
        - name: to
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Purchase register
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PurchaseRegisterRow"
  /reports/cash-book:
    get:
      operationId: getCashBook
      tags: [reports]
      summary: Cash book
      security:
        - bearerAuth: []
      parameters:
        - name: from
          in: query
          required: true
          schema:
            type: string
        - name: to
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Cash book
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CashBookRow"
  /reports/stock-current:
    get:
      operationId: getCurrentStock
      tags: [reports]
      summary: Current stock report
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Current stock
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CurrentStockRow"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    InvoiceItem:
      type: object
      properties:
        id:
          type: integer
        stockItemId:
          type: ["integer", "null"]
        itemName:
          type: string
        hsnCode:
          type: ["string", "null"]
        quantity:
          type: number
        unit:
          type: string
        rate:
          type: number
        discountPct:
          type: number
        gstPct:
          type: number
        taxableAmount:
          type: number
        cgst:
          type: number
        sgst:
          type: number
        igst:
          type: number
        total:
          type: number
      required:
        - id
        - itemName
        - quantity
        - unit
        - rate
        - discountPct
        - gstPct
        - taxableAmount
        - cgst
        - sgst
        - igst
        - total
    InvoiceItemInput:
      type: object
      properties:
        stockItemId:
          type: ["integer", "null"]
        itemName:
          type: string
        hsnCode:
          type: ["string", "null"]
        quantity:
          type: number
        unit:
          type: string
        rate:
          type: number
        discountPct:
          type: number
        gstPct:
          type: number
      required:
        - itemName
        - quantity
        - unit
        - rate
        - discountPct
        - gstPct
    PaymentMode:
      type: object
      properties:
        mode:
          type: string
          enum: [cash, upi, cheque, bank_transfer, credit]
        amount:
          type: number
        reference:
          type: ["string", "null"]
      required:
        - mode
        - amount
    SaleInvoice:
      type: object
      properties:
        id:
          type: integer
        invoiceNumber:
          type: string
        date:
          type: string
        partyId:
          type: ["integer", "null"]
        partyName:
          type: string
        partyGstin:
          type: ["string", "null"]
        billingAddress:
          type: ["string", "null"]
        isGst:
          type: boolean
        isInterstate:
          type: boolean
        subtotal:
          type: number
        totalDiscount:
          type: number
        totalTaxable:
          type: number
        totalCgst:
          type: number
        totalSgst:
          type: number
        totalIgst:
          type: number
        totalGst:
          type: number
        grandTotal:
          type: number
        amountPaid:
          type: number
        balanceDue:
          type: number
        notes:
          type: ["string", "null"]
        status:
          type: string
          enum: [draft, confirmed, cancelled]
        items:
          type: array
          items:
            $ref: "#/components/schemas/InvoiceItem"
        payments:
          type: array
          items:
            $ref: "#/components/schemas/PaymentMode"
        createdAt:
          type: string
      required:
        - id
        - invoiceNumber
        - date
        - partyName
        - isGst
        - isInterstate
        - subtotal
        - totalDiscount
        - totalTaxable
        - totalCgst
        - totalSgst
        - totalIgst
        - totalGst
        - grandTotal
        - amountPaid
        - balanceDue
        - status
        - items
        - payments
        - createdAt
    CreateSaleInvoiceBody:
      type: object
      properties:
        date:
          type: string
        partyId:
          type: ["integer", "null"]
        partyName:
          type: string
        partyGstin:
          type: ["string", "null"]
        billingAddress:
          type: ["string", "null"]
        isGst:
          type: boolean
        isInterstate:
          type: boolean
        notes:
          type: ["string", "null"]
        items:
          type: array
          items:
            $ref: "#/components/schemas/InvoiceItemInput"
        payments:
          type: array
          items:
            $ref: "#/components/schemas/PaymentMode"
      required:
        - date
        - partyName
        - isGst
        - isInterstate
        - items
        - payments
    UpdateSaleInvoiceBody:
      type: object
      properties:
        date:
          type: string
        partyId:
          type: ["integer", "null"]
        partyName:
          type: string
        partyGstin:
          type: ["string", "null"]
        billingAddress:
          type: ["string", "null"]
        isGst:
          type: boolean
        isInterstate:
          type: boolean
        notes:
          type: ["string", "null"]
        items:
          type: array
          items:
            $ref: "#/components/schemas/InvoiceItemInput"
        payments:
          type: array
          items:
            $ref: "#/components/schemas/PaymentMode"
        status:
          type: string
          enum: [draft, confirmed, cancelled]
...
    DayBookEntry:
      type: object
      properties:
        id:
          type: integer
        time:
          type: string
        voucherType:
          type: string
        voucherNumber:
          type: string
        partyName:
          type: ["string", "null"]
        narration:
          type: string
        debit:
          type: ["number", "null"]
        credit:
          type: ["number", "null"]
        amount:
          type: number
      required:
        - id
        - time
        - voucherType
        - voucherNumber
        - narration
        - amount
    TrialBalanceRow:
      type: object
      properties:
        ledgerId:
          type: integer
        ledgerName:
          type: string
        group:
          type: string
        openingDr:
          type: number
        openingCr:
          type: number
        transactionDr:
          type: number
        transactionCr:
          type: number
        closingDr:
          type: number
        closingCr:
          type: number
      required:
        - ledgerId
        - ledgerName
        - group
        - openingDr
        - openingCr
        - transactionDr
        - transactionCr
        - closingDr
        - closingCr
    ProfitLoss:
      type: object
      properties:
        grossIncome:
          type: number
        grossExpense:
          type: number
        grossProfit:
          type: number
        netProfit:
          type: number
        incomeGroups:
          type: array
          items:
            $ref: "#/components/schemas/TrialBalanceRow"
        expenseGroups:
          type: array
          items:
            $ref: "#/components/schemas/TrialBalanceRow"
      required:
        - grossIncome
        - grossExpense
        - grossProfit
        - netProfit
        - incomeGroups
        - expenseGroups
    BalanceSheet:
      type: object
      properties:
        totalAssets:
          type: number
        totalLiabilities:
          type: number
        assets:
          type: array
          items:
            $ref: "#/components/schemas/TrialBalanceRow"
        liabilities:
          type: array
          items:
            $ref: "#/components/schemas/TrialBalanceRow"
      required:
        - totalAssets
        - totalLiabilities
        - assets
        - liabilities
    SaleRegisterRow:
      type: object
      properties:
        id:
          type: integer
        invoiceNumber:
          type: string
        date:
          type: string
        partyName:
          type: string
        taxableAmount:
          type: number
        cgst:
          type: number
        sgst:
          type: number
        igst:
          type: number
        totalGst:
          type: number
        grandTotal:
          type: number
        balanceDue:
          type: number
      required:
        - id
        - invoiceNumber
        - date
        - partyName
        - taxableAmount
        - cgst
        - sgst
        - igst
        - totalGst
        - grandTotal
        - balanceDue
    PurchaseRegisterRow:
      type: object
      properties:
        id:
          type: integer
        invoiceNumber:
          type: string
        date:
          type: string
        partyName:
          type: string
        taxableAmount:
          type: number
        cgst:
          type: number
        sgst:
          type: number
        igst:
          type: number
        totalGst:
          type: number
        grandTotal:
          type: number
      required:
        - id
        - invoiceNumber
        - date
        - partyName
        - taxableAmount
        - cgst
        - sgst
        - igst
        - totalGst
        - grandTotal
    CashBookRow:
      type: object
      properties:
        id:
          type: integer
        date:
          type: string
        voucherType:
          type: string
        voucherNumber:
          type: string
        narration:
          type: string
        debit:
          type: ["number", "null"]
        credit:
          type: ["number", "null"]
        balance:
          type: number
      required:
        - id
        - date
        - voucherType
        - voucherNumber
        - narration
        - balance
    CurrentStockRow:
      type: object
      properties:
        id:
          type: integer
        itemName:
          type: string
        categoryName:
          type: ["string", "null"]
        unit:
          type: string
        physicalStock:
          type: number
        availableStock:
          type: number
        reservedStock:
          type: number
        minStockLevel:
          type: number
        purchaseRate:
          type: number
        saleRate:
          type: number
        stockValue:
          type: number
        isLow:
          type: boolean
      required:
        - id
        - itemName
        - unit
        - physicalStock
        - availableStock
        - reservedStock
        - minStockLevel
        - purchaseRate
        - saleRate
        - stockValue
        - isLow